save project OD preference in Maven projects too
