+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject/META-INF/MANIFEST.MF
@@ -20,6 +20,7 @@ Require-Bundle: com.google.cloud.tools.appengine;bundle-version="0.1.0",
  com.google.cloud.tools.eclipse.sdk.ui;bundle-version="0.1.0"
 Import-Package: com.google.cloud.tools.eclipse.appengine.facets,
  com.google.cloud.tools.eclipse.appengine.libraries,
+ com.google.cloud.tools.eclipse.preferences,
  com.google.cloud.tools.eclipse.usagetracker,
  com.google.cloud.tools.eclipse.util.templates.appengine,
  com.google.common.annotations;version="15.0.0",
